[add] - usequery파일 옮김
